requirements.txt
pandas
numpy
scikit-learn
xgboost
boto3
joblib
python-dotenv==1.0.0 